frontend/refactor: file picker
